CRAN-like Repository Structure

CRAN-like repository structure for packages in an R-universe.

runiv_src(universe)

runiv_bin(universe)

Arguments

universe
A length one character. User or organization name of the the R-universe owner.

Value

a data.frame.

Author

Valerio Gherardi

Examples

runiv_src("vgherard")

#>                        _id  Package    Version    License NeedsCompilation
#> 1 60fd8124c6cd580b83b1a3cd    runiv 0.0.0.9000 GPL (>= 3)               no
#> 2 60e0304de8525d5b6388c1e8      r2r      0.1.1 GPL (>= 3)               no
#> 3 60e02dfbe8525d5b6388bf01   kgrams      0.1.0 GPL (>= 3)              yes
#> 4 60e02c64e8525d5b6388bd4c scribblr 0.2.0.9000 GPL (>= 3)               no
#> 5 60e02c2de8525d5b6388bcb7  gsample      0.1.0      GPL-3              yes
#> 6 60e02522e8525d5b6388b942      sbo      0.5.0      GPL-3              yes
#> 7 60e02362e8525d5b6388b85a     fcci      1.0.0 GPL (>= 3)              yes
#>                             MD5sum
#> 1 3aa0d2de995ed2a318fbf9165d0974a3
#> 2 6250f28963e5cf34914a83c4a290bad1
#> 3 c20d7103b32574373cb0a576e3be6c96
#> 4 9bc8757ade7e344ee9b1c955dd3b7fd3
#> 5 8340aad7192cd6b7c5d04565e9ac00af
#> 6 acd1abeb9b46a1c607d9f60aea0b8510
#> 7 9ca24cd15c82ac543d5f205096d5eba4
#>                                                                                                                                                                                                                                            _hard_deps
#> 1                                                                                                                                                                                     curl, httr, jsonlite, rlang, Imports, Imports, Imports, Imports
#> 2                                                                                                                                                                                                                     digest, rlang, Imports, Imports
#> 3                                           Rcpp, RcppProgress, Rcpp, rlang, methods, utils, RcppProgress, Rdpack, R, LinkingTo, LinkingTo, Imports, Imports, Imports, Imports, Imports, Imports, Depends, NA, NA, NA, NA, NA, NA, >= 0.1, NA, >= 3.5
#> 4 shiny, miniUI, shinyjs, rstudioapi, fs, usethis, keys, gh, shinyWidgets, utils, rlang, >= 0.13, >= 0.1.1, NA, >= 0.5, NA, NA, NA, NA, NA, NA, NA, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports
#> 5                                                                                                                                                                                                                      Rcpp, Rcpp, LinkingTo, Imports
#> 6                                    R, Rcpp, testthat, Rcpp, rlang, tidyr, dplyr, utils, stats, graphics, >= 3.5.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, Depends, LinkingTo, LinkingTo, Imports, Imports, Imports, Imports, Imports, Imports, Imports
#> 7                                                                                                                                                                                 Rcpp, Rcpp, assertthat, rlang, LinkingTo, Imports, Imports, Imports
#>                                                                                                                                         _soft_deps
#> 1                                                                                                 covr, testthat, Suggests, Suggests, NA, >= 3.0.0
#> 2 knitr, covr, testthat, rmarkdown, microbenchmark, hash, Suggests, Suggests, Suggests, Suggests, Suggests, Suggests, NA, NA, >= 3.0.0, NA, NA, NA
#> 3                                                   testthat, covr, knitr, rmarkdown, >= 3.0.0, NA, NA, NA, Suggests, Suggests, Suggests, Suggests
#> 4                                                                                                 covr, testthat, Suggests, Suggests, NA, >= 3.0.0
#> 5                                                                                         knitr, rmarkdown, testthat, Suggests, Suggests, Suggests
#> 6                                       ggplot2, knitr, rmarkdown, cli, testthat, covr, Suggests, Suggests, Suggests, Suggests, Suggests, Suggests
#> 7                                                                                                 covr, testthat, Suggests, Suggests, NA, >= 3.0.0