Packages in an R-universe

List all packages available in an R-universe.

runiv_packages(universe, package = NULL, version = NULL)

Arguments

universe
A length one character. User or organization name of the the R-universe owner.

package
either NULL, or a length one character. Package name (see details).

version
either NULL, or a character vector. Package version (see details).

Value

a data.frame if both package and version are not NULL, a character vector otherwise.

Details

This function gives access to three types of information, depending on how many arguments are different from NULL:

runiv_packages(universe) lists all packages in the R-universe universe.

runiv_packages(universe, package), where package is a string, lists all available version for package package in this universe.

runiv_packages(universe, package, version) lists all available builds for package package of version version (specified as a string).

The argument version is ignored if package is NULL.

Author

Valerio Gherardi

Examples

( packages <- runiv_packages("vgherard") )

#> [1] "fcci"     "gsample"  "kgrams"   "r2r"      "runiv"    "sbo"      "scribblr"

( versions <- runiv_packages("vgherard", packages[[1]]) )

#> [1] "1.0.0"

( versions <- runiv_packages("vgherard", packages[[1]], versions[[1]]) )

#>                        _id Package    Type                                Title
#> 1 60e02362e8525d5b6388b84a    fcci Package Feldman-Cousins Confidence Intervals
#> 2 60e02362e8525d5b6388b852    fcci Package Feldman-Cousins Confidence Intervals
#> 3 60e02362e8525d5b6388b85a    fcci Package Feldman-Cousins Confidence Intervals
#> 4 60e02362e8525d5b6388b864    fcci Package Feldman-Cousins Confidence Intervals
#> 5 60e02365e8525d5b6388b86f    fcci Package Feldman-Cousins Confidence Intervals
#>   Version
#> 1   1.0.0
#> 2   1.0.0
#> 3   1.0.0
#> 4   1.0.0
#> 5   1.0.0
#>                                                                                                                                             Authors@R
#> 1 person(given = "Valerio",\nfamily = "Gherardi",\nrole = c("aut", "cre"),\nemail = "vgherard@sissa.it",\ncomment = c(ORCID = "0000-0002-8215-3013"))
#> 2 person(given = "Valerio",\nfamily = "Gherardi",\nrole = c("aut", "cre"),\nemail = "vgherard@sissa.it",\ncomment = c(ORCID = "0000-0002-8215-3013"))
#> 3 person(given = "Valerio",\nfamily = "Gherardi",\nrole = c("aut", "cre"),\nemail = "vgherard@sissa.it",\ncomment = c(ORCID = "0000-0002-8215-3013"))
#> 4 person(given = "Valerio",\nfamily = "Gherardi",\nrole = c("aut", "cre"),\nemail = "vgherard@sissa.it",\ncomment = c(ORCID = "0000-0002-8215-3013"))
#> 5 person(given = "Valerio",\nfamily = "Gherardi",\nrole = c("aut", "cre"),\nemail = "vgherard@sissa.it",\ncomment = c(ORCID = "0000-0002-8215-3013"))
#>                             Maintainer
#> 1 Valerio Gherardi <vgherard@sissa.it>
#> 2 Valerio Gherardi <vgherard@sissa.it>
#> 3 Valerio Gherardi <vgherard@sissa.it>
#> 4 Valerio Gherardi <vgherard@sissa.it>
#> 5 Valerio Gherardi <vgherard@sissa.it>
#>                                                                                                                                    Description
#> 1 Provides support for building Feldman-Cousins confidence intervals\n[G. J. Feldman and R. D. Cousins (1998) <doi:10.1103/PhysRevD.57.3873>].
#> 2 Provides support for building Feldman-Cousins confidence intervals\n[G. J. Feldman and R. D. Cousins (1998) <doi:10.1103/PhysRevD.57.3873>].
#> 3 Provides support for building Feldman-Cousins confidence intervals\n[G. J. Feldman and R. D. Cousins (1998) <doi:10.1103/PhysRevD.57.3873>].
#> 4 Provides support for building Feldman-Cousins confidence intervals\n[G. J. Feldman and R. D. Cousins (1998) <doi:10.1103/PhysRevD.57.3873>].
#> 5 Provides support for building Feldman-Cousins confidence intervals\n[G. J. Feldman and R. D. Cousins (1998) <doi:10.1103/PhysRevD.57.3873>].
#>      License Encoding RoxygenNote SystemRequirements
#> 1 GPL (>= 3)    UTF-8       7.1.1              C++11
#> 2 GPL (>= 3)    UTF-8       7.1.1              C++11
#> 3 GPL (>= 3)    UTF-8       7.1.1              C++11
#> 4 GPL (>= 3)    UTF-8       7.1.1              C++11
#> 5 GPL (>= 3)    UTF-8       7.1.1              C++11
#>                                URL                              BugReports
#> 1 https://github.com/vgherard/fcci https://github.com/vgherard/fcci/issues
#> 2 https://github.com/vgherard/fcci https://github.com/vgherard/fcci/issues
#> 3 https://github.com/vgherard/fcci https://github.com/vgherard/fcci/issues
#> 4 https://github.com/vgherard/fcci https://github.com/vgherard/fcci/issues
#> 5 https://github.com/vgherard/fcci https://github.com/vgherard/fcci/issues
#>   Config/testthat/edition                      Repository
#> 1                       3 https://vgherard.r-universe.dev
#> 2                       3 https://vgherard.r-universe.dev
#> 3                       3 https://vgherard.r-universe.dev
#> 4                       3 https://vgherard.r-universe.dev
#> 5                       3 https://vgherard.r-universe.dev
#>                          RemoteUrl RemoteRef
#> 1 https://github.com/vgherard/fcci    v1.0.0
#> 2 https://github.com/vgherard/fcci    v1.0.0
#> 3 https://github.com/vgherard/fcci    v1.0.0
#> 4 https://github.com/vgherard/fcci    v1.0.0
#> 5 https://github.com/vgherard/fcci    v1.0.0
#>                                  RemoteSha NeedsCompilation
#> 1 3df3ccab1ab9f0880c9a4140430851ff366a3669              yes
#> 2 3df3ccab1ab9f0880c9a4140430851ff366a3669              yes
#> 3 3df3ccab1ab9f0880c9a4140430851ff366a3669              yes
#> 4 3df3ccab1ab9f0880c9a4140430851ff366a3669              yes
#> 5 3df3ccab1ab9f0880c9a4140430851ff366a3669              yes
#>                        Packaged
#> 1 2021-07-03 09:16:24 UTC; root
#> 2 2021-07-03 09:16:24 UTC; root
#> 3 2021-07-03 09:16:24 UTC; root
#> 4 2021-07-03 09:16:24 UTC; root
#> 5 2021-07-03 09:16:24 UTC; root
#>                                                                  Author Built.R
#> 1 Valerio Gherardi [aut, cre] (<https://orcid.org/0000-0002-8215-3013>)   4.0.5
#> 2 Valerio Gherardi [aut, cre] (<https://orcid.org/0000-0002-8215-3013>)   4.1.0
#> 3 Valerio Gherardi [aut, cre] (<https://orcid.org/0000-0002-8215-3013>)    <NA>
#> 4 Valerio Gherardi [aut, cre] (<https://orcid.org/0000-0002-8215-3013>)   4.0.5
#> 5 Valerio Gherardi [aut, cre] (<https://orcid.org/0000-0002-8215-3013>)   4.1.0
#>            Built.Platform              Built.Date Built.OStype    _user _type
#> 1 x86_64-apple-darwin17.0 2021-07-03 09:22:46 UTC         unix vgherard   mac
#> 2 x86_64-apple-darwin17.0 2021-07-03 09:19:46 UTC         unix vgherard   mac
#> 3                    <NA>                    <NA>         <NA> vgherard   src
#> 4      x86_64-w64-mingw32 2021-07-03 09:17:56 UTC      windows vgherard   win
#> 5      x86_64-w64-mingw32 2021-07-03 09:18:22 UTC      windows vgherard   win
#>               _file               _published                _builder.upstream
#> 1    fcci_1.0.0.tgz 2021-07-03T09:23:40.826Z https://github.com/vgherard/fcci
#> 2    fcci_1.0.0.tgz 2021-07-03T09:23:41.044Z https://github.com/vgherard/fcci
#> 3 fcci_1.0.0.tar.gz 2021-07-03T09:23:41.181Z https://github.com/vgherard/fcci
#> 4    fcci_1.0.0.zip 2021-07-03T09:23:41.345Z https://github.com/vgherard/fcci
#> 5    fcci_1.0.0.zip 2021-07-03T09:23:41.635Z https://github.com/vgherard/fcci
#>   _builder.date                          _builder.commit _builder.registered
#> 1    1625304220 3df3ccab1ab9f0880c9a4140430851ff366a3669                true
#> 2    1625304220 3df3ccab1ab9f0880c9a4140430851ff366a3669                true
#> 3    1625304221 3df3ccab1ab9f0880c9a4140430851ff366a3669                true
#> 4    1625304221 3df3ccab1ab9f0880c9a4140430851ff366a3669                true
#> 5    1625304221 3df3ccab1ab9f0880c9a4140430851ff366a3669                true
#>   _builder.timestamp _builder.maintainerlogin _builder.distro  _builder.host
#> 1         1617373968                 vgherard           macos GitHub-Actions
#> 2         1617373968                 vgherard           macos GitHub-Actions
#> 3         1617373968                 vgherard           focal GitHub-Actions
#> 4         1617373968                 vgherard         windows GitHub-Actions
#> 5         1617373968                 vgherard         windows GitHub-Actions
#>   _builder.status                                                  _builder.url
#> 1         success https://github.com/r-universe/vgherard/actions/runs/996009318
#> 2         success https://github.com/r-universe/vgherard/actions/runs/996009318
#> 3         success https://github.com/r-universe/vgherard/actions/runs/996009318
#> 4         success https://github.com/r-universe/vgherard/actions/runs/996009318
#> 5         success https://github.com/r-universe/vgherard/actions/runs/996009318
#>                    _builder.sysdeps                           MD5sum
#> 1                              NULL 6c24c681dd48178eb515a224b00dd00f
#> 2                              NULL 0ffbb562ace26bde3670e5a48f5463d0
#> 3 libstdc++6, 10.2.0-5ubuntu1~20.04 9ca24cd15c82ac543d5f205096d5eba4
#> 4                              NULL f641390c413a4c2d120edebdca862671
#> 5                              NULL d40b1f82490e2d126ffba08dd9f4d760
#>                                                            _hard_deps
#> 1 Rcpp, Rcpp, assertthat, rlang, LinkingTo, Imports, Imports, Imports
#> 2 Rcpp, Rcpp, assertthat, rlang, LinkingTo, Imports, Imports, Imports
#> 3 Rcpp, Rcpp, assertthat, rlang, LinkingTo, Imports, Imports, Imports
#> 4 Rcpp, Rcpp, assertthat, rlang, LinkingTo, Imports, Imports, Imports
#> 5 Rcpp, Rcpp, assertthat, rlang, LinkingTo, Imports, Imports, Imports
#>                                         _soft_deps     Archs
#> 1 covr, testthat, Suggests, Suggests, NA, >= 3.0.0      <NA>
#> 2 covr, testthat, Suggests, Suggests, NA, >= 3.0.0      <NA>
#> 3 covr, testthat, Suggests, Suggests, NA, >= 3.0.0      <NA>
#> 4 covr, testthat, Suggests, Suggests, NA, >= 3.0.0 i386, x64
#> 5 covr, testthat, Suggests, Suggests, NA, >= 3.0.0 i386, x64