Recent Builds and Checks

Get list with recent builds and checks of packages available in an R-universe.

runiv_stats_checks(universe)

Arguments

universe
A length one character. User or organization name of the the R-universe owner.

Value

a data.frame.

Author

Valerio Gherardi

Examples

runiv_stats_checks("vgherard")

#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     runs
#> 1                                                                                                                                                                                                                           win, mac, win, mac, src, https://github.com/vgherard/runiv, https://github.com/vgherard/runiv, https://github.com/vgherard/runiv, https://github.com/vgherard/runiv, https://github.com/vgherard/runiv, 1627396156, 1627396155, 1627396156, 1627396155, 1627396156, 0af8ece827100d23ffa2e1ed10608842cf1ddcaf, 0af8ece827100d23ffa2e1ed10608842cf1ddcaf, 0af8ece827100d23ffa2e1ed10608842cf1ddcaf, 0af8ece827100d23ffa2e1ed10608842cf1ddcaf, 0af8ece827100d23ffa2e1ed10608842cf1ddcaf, true, true, true, true, true, 1627394906, 1627394906, 1627394906, 1627394906, 1627394906, vgherard, vgherard, vgherard, vgherard, vgherard, windows, macos, windows, macos, focal, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, success, success, success, success, success, https://github.com/r-universe/vgherard/actions/runs/1071473639, https://github.com/r-universe/vgherard/actions/runs/1071473639, https://github.com/r-universe/vgherard/actions/runs/1071473639, https://github.com/r-universe/vgherard/actions/runs/1071473639, https://github.com/r-universe/vgherard/actions/runs/1071473639, 4.1.0, 4.0.5, 4.0.5, 4.1.0, NA, , , , , NA, 2021-07-27 14:21:38 UTC, 2021-07-27 14:25:23 UTC, 2021-07-27 14:20:11 UTC, 2021-07-27 14:28:40 UTC, NA, windows, unix, windows, unix, NA, 2021-07-27T14:29:16.530Z, 2021-07-27T14:29:15.854Z, 2021-07-27T14:29:16.372Z, 2021-07-27T14:29:16.016Z, 2021-07-27T14:29:16.214Z
#> 2                                                                                                                        src, win, mac, win, mac, https://github.com/vgherard/scribblr, https://github.com/vgherard/scribblr, https://github.com/vgherard/scribblr, https://github.com/vgherard/scribblr, https://github.com/vgherard/scribblr, 1626456496, 1626456497, 1626456495, 1626456496, 1626456496, 7ae46bf068533d6a9956d97ab89a69b2ba6471f4, 7ae46bf068533d6a9956d97ab89a69b2ba6471f4, 7ae46bf068533d6a9956d97ab89a69b2ba6471f4, 7ae46bf068533d6a9956d97ab89a69b2ba6471f4, 7ae46bf068533d6a9956d97ab89a69b2ba6471f4, true, true, true, true, true, 1626452731, 1626452731, 1626452731, 1626452731, 1626452731, vgherard, vgherard, vgherard, vgherard, vgherard, focal, windows, macos, windows, macos, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, success, success, success, success, success, https://github.com/vgherard/scribblr/raw/HEAD/man/figures/logo.png, NA, NA, NA, NA, https://github.com/r-universe/vgherard/actions/runs/1038259138, https://github.com/r-universe/vgherard/actions/runs/1038259138, https://github.com/r-universe/vgherard/actions/runs/1038259138, https://github.com/r-universe/vgherard/actions/runs/1038259138, https://github.com/r-universe/vgherard/actions/runs/1038259138, 2021-07-16T17:28:16.646Z, 2021-07-16T17:28:17.269Z, 2021-07-16T17:28:16.003Z, 2021-07-16T17:28:16.966Z, 2021-07-16T17:28:16.369Z, NA, 4.1.0, 4.0.5, 4.0.5, 4.1.0, NA, , , , , NA, 2021-07-16 17:26:36 UTC, 2021-07-16 17:27:33 UTC, 2021-07-16 17:25:58 UTC, 2021-07-16 17:26:38 UTC, NA, windows, unix, windows, unix
#> 3                                                                      win, src, mac, win, mac, https://github.com/vgherard/r2r, https://github.com/vgherard/r2r, https://github.com/vgherard/r2r, https://github.com/vgherard/r2r, https://github.com/vgherard/r2r, 1627226601, 1627226600, 1627226599, 1627226601, 1627226599, d9679b09071a7d71bb06924850d6566c8582e99b, d9679b09071a7d71bb06924850d6566c8582e99b, d9679b09071a7d71bb06924850d6566c8582e99b, d9679b09071a7d71bb06924850d6566c8582e99b, d9679b09071a7d71bb06924850d6566c8582e99b, true, true, true, true, true, 1625484226, 1625484226, 1625484226, 1625484226, 1625484226, vgherard, vgherard, vgherard, vgherard, vgherard, windows, focal, macos, windows, macos, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, success, success, success, success, success, https://github.com/r-universe/vgherard/actions/runs/1064845553, https://github.com/r-universe/vgherard/actions/runs/1064845553, https://github.com/r-universe/vgherard/actions/runs/1064845553, https://github.com/r-universe/vgherard/actions/runs/1064845553, https://github.com/r-universe/vgherard/actions/runs/1064845553, benchmarks.Rmd, r2r.Rmd, benchmarks.html, r2r.html, Comparison, r2r, 2021-07-02 11:17:51, 2021-07-01 15:23:59, 2021-07-02 12:04:10, 2021-07-02 11:17:51, 3, 2, 4.0.5, NA, 4.0.5, 4.1.0, 4.1.0, , NA, , , , 2021-07-25 15:21:41 UTC, NA, 2021-07-25 15:22:37 UTC, 2021-07-25 15:22:15 UTC, 2021-07-25 15:22:34 UTC, windows, NA, unix, windows, unix, 2021-07-25T15:23:21.530Z, 2021-07-25T15:23:20.981Z, 2021-07-25T15:23:19.893Z, 2021-07-25T15:23:22.412Z, 2021-07-25T15:23:20.444Z
#> 4                                                                                                                win, win, mac, mac, src, https://github.com/vgherard/fcci, https://github.com/vgherard/fcci, https://github.com/vgherard/fcci, https://github.com/vgherard/fcci, https://github.com/vgherard/fcci, 1625304221, 1625304221, 1625304220, 1625304220, 1625304221, 3df3ccab1ab9f0880c9a4140430851ff366a3669, 3df3ccab1ab9f0880c9a4140430851ff366a3669, 3df3ccab1ab9f0880c9a4140430851ff366a3669, 3df3ccab1ab9f0880c9a4140430851ff366a3669, 3df3ccab1ab9f0880c9a4140430851ff366a3669, true, true, true, true, true, 1617373968, 1617373968, 1617373968, 1617373968, 1617373968, vgherard, vgherard, vgherard, vgherard, vgherard, windows, windows, macos, macos, focal, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, success, success, success, success, success, https://github.com/r-universe/vgherard/actions/runs/996009318, https://github.com/r-universe/vgherard/actions/runs/996009318, https://github.com/r-universe/vgherard/actions/runs/996009318, https://github.com/r-universe/vgherard/actions/runs/996009318, https://github.com/r-universe/vgherard/actions/runs/996009318, libstdc++6, 10.2.0-5ubuntu1~20.04, 4.0.5, 4.1.0, 4.0.5, 4.1.0, NA, x86_64-w64-mingw32, x86_64-w64-mingw32, x86_64-apple-darwin17.0, x86_64-apple-darwin17.0, NA, 2021-07-03 09:17:56 UTC, 2021-07-03 09:18:22 UTC, 2021-07-03 09:22:46 UTC, 2021-07-03 09:19:46 UTC, NA, windows, windows, unix, unix, NA, 2021-07-03T09:23:41.345Z, 2021-07-03T09:23:41.635Z, 2021-07-03T09:23:40.826Z, 2021-07-03T09:23:41.044Z, 2021-07-03T09:23:41.181Z
#> 5 win, mac, mac, src, win, https://github.com/vgherard/kgrams, https://github.com/vgherard/kgrams, https://github.com/vgherard/kgrams, https://github.com/vgherard/kgrams, https://github.com/vgherard/kgrams, 1625981187, 1625981186, 1625981186, 1625981187, 1625981187, 3760408b9d0eafd4ff826201c9a5576f57a25500, 3760408b9d0eafd4ff826201c9a5576f57a25500, 3760408b9d0eafd4ff826201c9a5576f57a25500, 3760408b9d0eafd4ff826201c9a5576f57a25500, 3760408b9d0eafd4ff826201c9a5576f57a25500, true, true, true, true, true, 1613042037, 1613042037, 1613042037, 1613042037, 1613042037, vgherard, vgherard, vgherard, vgherard, vgherard, windows, macos, macos, focal, windows, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, success, success, success, success, success, https://github.com/r-universe/vgherard/actions/runs/1019368977, https://github.com/r-universe/vgherard/actions/runs/1019368977, https://github.com/r-universe/vgherard/actions/runs/1019368977, https://github.com/r-universe/vgherard/actions/runs/1019368977, https://github.com/r-universe/vgherard/actions/runs/1019368977, kgrams.Rmd, kgrams.html, kgrams, Valerio Gherardi, 2021-02-06 18:04:34, 2021-02-11 11:13:57, 4, libstdc++6, 10.2.0-5ubuntu1~20.04, 4.1.0, 4.1.0, 4.0.5, NA, 4.0.5, x86_64-w64-mingw32, x86_64-apple-darwin17.0, x86_64-apple-darwin17.0, NA, x86_64-w64-mingw32, 2021-07-11 05:21:37 UTC, 2021-07-11 05:20:00 UTC, 2021-07-11 05:01:46 UTC, NA, 2021-07-11 05:08:51 UTC, windows, unix, unix, NA, windows, 2021-07-11T05:26:27.827Z, 2021-07-11T05:26:27.129Z, 2021-07-11T05:26:26.929Z, 2021-07-11T05:26:27.299Z, 2021-07-11T05:26:27.568Z
#> 6                                                                                            win, win, src, mac, mac, https://github.com/vgherard/gsample, https://github.com/vgherard/gsample, https://github.com/vgherard/gsample, https://github.com/vgherard/gsample, https://github.com/vgherard/gsample, 1626673803, 1626673802, 1626673802, 1626673799, 1626673801, 5b1bbc8e768417389b6ca1451e80f93457cbcfba, 5b1bbc8e768417389b6ca1451e80f93457cbcfba, 5b1bbc8e768417389b6ca1451e80f93457cbcfba, 5b1bbc8e768417389b6ca1451e80f93457cbcfba, 5b1bbc8e768417389b6ca1451e80f93457cbcfba, true, true, true, true, true, 1611175216, 1611175216, 1611175216, 1611175216, 1611175216, vgherard, vgherard, vgherard, vgherard, vgherard, windows, windows, focal, macos, macos, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, failure, failure, success, failure, failure, https://github.com/r-universe/vgherard/actions/runs/1043988360, https://github.com/r-universe/vgherard/actions/runs/1043988360, https://github.com/r-universe/vgherard/actions/runs/1043988360, https://github.com/r-universe/vgherard/actions/runs/1043988360, https://github.com/r-universe/vgherard/actions/runs/1043988360, libstdc++6, 10.2.0-5ubuntu1~20.04, 4.1.0, 4.0.5, NA, 4.0.5, 4.1.0, x86_64-w64-mingw32, x86_64-w64-mingw32, NA, x86_64-apple-darwin17.0, x86_64-apple-darwin17.0, 2021-07-19 05:22:15 UTC, 2021-07-19 05:14:24 UTC, NA, 2021-07-19 05:47:49 UTC, 2021-07-19 05:28:20 UTC, windows, windows, NA, unix, unix, 2021-07-19T05:50:04.123Z, 2021-07-19T05:50:03.336Z, 2021-07-19T05:50:02.640Z, 2021-07-19T05:50:01.370Z, 2021-07-19T05:50:02.110Z
#> 7                             mac, win, win, src, mac, https://github.com/vgherard/sbo, https://github.com/vgherard/sbo, https://github.com/vgherard/sbo, https://github.com/vgherard/sbo, https://github.com/vgherard/sbo, 1625302305, 1625302357, 1625302306, 1625302306, 1625302305, 75374a5bf57709fae780207fa1f93b8e86d7c6c4, 75374a5bf57709fae780207fa1f93b8e86d7c6c4, 75374a5bf57709fae780207fa1f93b8e86d7c6c4, 75374a5bf57709fae780207fa1f93b8e86d7c6c4, 75374a5bf57709fae780207fa1f93b8e86d7c6c4, true, true, true, true, true, 1607180351, 1607180351, 1607180351, 1607180351, 1607180351, vgherard, vgherard, vgherard, vgherard, vgherard, macos, windows, windows, focal, macos, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, GitHub-Actions, success, success, success, success, success, https://github.com/r-universe/vgherard/actions/runs/995955933, https://github.com/r-universe/vgherard/actions/runs/995955933, https://github.com/r-universe/vgherard/actions/runs/995955933, https://github.com/r-universe/vgherard/actions/runs/995955933, https://github.com/r-universe/vgherard/actions/runs/995955933, sbo.Rmd, sbo.html, sbo, Valerio Gherardi, 2020-08-01 22:17:39, 2020-12-05 11:01:42, 10, libstdc++6, 10.2.0-5ubuntu1~20.04, 4.1.0, 4.1.0, 4.0.5, NA, 4.0.5, x86_64-apple-darwin17.0, x86_64-w64-mingw32, x86_64-w64-mingw32, NA, x86_64-apple-darwin17.0, 2021-07-03 08:49:21 UTC, 2021-07-03 08:45:19 UTC, 2021-07-03 08:44:54 UTC, NA, 2021-07-03 08:46:02 UTC, unix, windows, windows, NA, unix, 2021-07-03T08:51:46.363Z, 2021-07-03T08:52:38.010Z, 2021-07-03T08:52:37.255Z, 2021-07-03T08:51:46.882Z, 2021-07-03T08:51:45.619Z
#>       user                           maintainer  package    version
#> 1 vgherard Valerio Gherardi <vgherard@sissa.it>    runiv 0.0.0.9000
#> 2 vgherard Valerio Gherardi <vgherard@sissa.it> scribblr 0.2.0.9000
#> 3 vgherard Valerio Gherardi <vgherard@sissa.it>      r2r      0.1.1
#> 4 vgherard Valerio Gherardi <vgherard@sissa.it>     fcci      1.0.0
#> 5 vgherard Valerio Gherardi <vgherard@sissa.it>   kgrams      0.1.0
#> 6 vgherard Valerio Gherardi <vgherard@sissa.it>  gsample      0.1.0
#> 7 vgherard Valerio Gherardi <vgherard@sissa.it>      sbo      0.5.0