DESCRIPTIONs

Get DESCRIPTION files of packages available in an R-universe.

runiv_stats_descriptions(universe)

Arguments

universe
A length one character. User or organization name of the the R-universe owner.

Value

a data.frame.

Author

Valerio Gherardi

Examples

runiv_stats_descriptions("vgherard")

#>                        _id  Package
#> 1 60fd8124c6cd580b83b1a3cd    runiv
#> 2 60e0304de8525d5b6388c1e8      r2r
#> 3 60e02dfbe8525d5b6388bf01   kgrams
#> 4 60e02c64e8525d5b6388bd4c scribblr
#> 5 60e02c2de8525d5b6388bcb7  gsample
#> 6 60e02522e8525d5b6388b942      sbo
#> 7 60e02362e8525d5b6388b85a     fcci
#>                                               Title    Version
#> 1                        Client for R-universe APIs 0.0.0.9000
#> 2                    R-Object to R-Object Hash Maps      0.1.1
#> 3                  Classical k-gram Language Models      0.1.0
#> 4                          A Notepad Inside RStudio 0.2.0.9000
#> 5   Efficient Weighted Sampling Without Replacement      0.1.0
#> 6 Text Prediction via Stupid Back-Off N-Gram Models      0.5.0
#> 7              Feldman-Cousins Confidence Intervals      1.0.0
#>                                                                                                                                             Authors@R
#> 1 person(given = "Valerio",\nfamily = "Gherardi",\nrole = c("aut", "cre"),\nemail = "vgherard@sissa.it",\ncomment = c(ORCID = "0000-0002-8215-3013"))
#> 2 person(given = "Valerio",\nfamily = "Gherardi",\nrole = c("aut", "cre"),\nemail = "vgherard@sissa.it",\ncomment = c(ORCID = "0000-0002-8215-3013"))
#> 3 person(given = "Valerio",\nfamily = "Gherardi",\nrole = c("aut", "cre"),\nemail = "vgherard@sissa.it",\ncomment = c(ORCID = "0000-0002-8215-3013"))
#> 4 person(given = "Valerio",\nfamily = "Gherardi",\nrole = c("aut", "cre"),\nemail = "vgherard@sissa.it",\ncomment = c(ORCID = "0000-0002-8215-3013"))
#> 5                                                                                                                                                <NA>
#> 6                                                                                                                                                <NA>
#> 7 person(given = "Valerio",\nfamily = "Gherardi",\nrole = c("aut", "cre"),\nemail = "vgherard@sissa.it",\ncomment = c(ORCID = "0000-0002-8215-3013"))
#>                                                                                                                                                                                                     Description
#> 1                                                                                                                                                                                   Client for R-universe APIs.
#> 2 Implementation of hash tables (hash sets and hash maps) in R,\nfeaturing arbitrary R objects as keys,\narbitrary hash and key-comparison functions,\nand customizable behaviour upon queries of missing keys.
#> 3                               Tools for training and evaluating k-gram language models in R,\nsupporting several probability smoothing techniques,\nperplexity computations, random text generation and more.
#> 4                                                                                           A project aware notepad inside RStudio, for taking quick project-related notes without distractions. RStudio addin.
#> 5                                                                                                     Sample without replacement using the Gumbel-Max trick (c.f. \\url{https://arxiv.org/pdf/1903.06059.pdf}).
#> 6                                             Utilities for training and evaluating text predictors based on Stupid Back-Off N-gram models (Brants et al., 2007, <https://www.aclweb.org/anthology/D07-1090/>).
#> 7                                                                  Provides support for building Feldman-Cousins confidence intervals\n[G. J. Feldman and R. D. Cousins (1998) <doi:10.1103/PhysRevD.57.3873>].
#>      License Encoding               Roxygen RoxygenNote Config/testthat/edition
#> 1 GPL (>= 3)    UTF-8 list(markdown = TRUE)       7.1.1                       3
#> 2 GPL (>= 3)    UTF-8 list(markdown = TRUE)       7.1.1                       3
#> 3 GPL (>= 3)    UTF-8 list(markdown = TRUE)       7.1.1                       3
#> 4 GPL (>= 3)    UTF-8 list(markdown = TRUE)       7.1.1                       3
#> 5      GPL-3    UTF-8                  <NA>       7.1.1                    <NA>
#> 6      GPL-3    UTF-8                  <NA>  7.1.1.9000                    <NA>
#> 7 GPL (>= 3)    UTF-8                  <NA>       7.1.1                       3
#>                                                                URL
#> 1                                https://github.com/vgherard/runiv
#> 2                                  https://github.com/vgherard/r2r
#> 3                                                             <NA>
#> 4                                                             <NA>
#> 5                              https://github.com/vgherard/gsample
#> 6 https://vgherard.github.io/sbo/, https://github.com/vgherard/sbo
#> 7                                 https://github.com/vgherard/fcci
#>                                    BugReports                      Repository
#> 1    https://github.com/vgherard/runiv/issues https://vgherard.r-universe.dev
#> 2      https://github.com/vgherard/r2r/issues https://vgherard.r-universe.dev
#> 3                                        <NA> https://vgherard.r-universe.dev
#> 4 https://github.com/vgherard/scribblr/issues https://vgherard.r-universe.dev
#> 5  https://github.com/vgherard/gsample/issues https://vgherard.r-universe.dev
#> 6      https://github.com/vgherard/sbo/issues https://vgherard.r-universe.dev
#> 7     https://github.com/vgherard/fcci/issues https://vgherard.r-universe.dev
#>                              RemoteUrl RemoteRef
#> 1    https://github.com/vgherard/runiv      HEAD
#> 2      https://github.com/vgherard/r2r    v0.1.1
#> 3   https://github.com/vgherard/kgrams    v0.1.0
#> 4 https://github.com/vgherard/scribblr      HEAD
#> 5  https://github.com/vgherard/gsample      HEAD
#> 6      https://github.com/vgherard/sbo      HEAD
#> 7     https://github.com/vgherard/fcci    v1.0.0
#>                                  RemoteSha NeedsCompilation
#> 1 00e5838d42c0ff8ccdd25342332ccd903bd6b765               no
#> 2 d9679b09071a7d71bb06924850d6566c8582e99b               no
#> 3 3760408b9d0eafd4ff826201c9a5576f57a25500              yes
#> 4 7ae46bf068533d6a9956d97ab89a69b2ba6471f4               no
#> 5 5b1bbc8e768417389b6ca1451e80f93457cbcfba              yes
#> 6 75374a5bf57709fae780207fa1f93b8e86d7c6c4              yes
#> 7 3df3ccab1ab9f0880c9a4140430851ff366a3669              yes
#>                        Packaged
#> 1 2021-07-27 15:18:10 UTC; root
#> 2 2021-07-25 15:20:20 UTC; root
#> 3 2021-07-11 04:57:42 UTC; root
#> 4 2021-07-16 17:24:14 UTC; root
#> 5 2021-07-19 04:58:27 UTC; root
#> 6 2021-07-03 08:42:03 UTC; root
#> 7 2021-07-03 09:16:24 UTC; root
#>                                                                  Author
#> 1 Valerio Gherardi [aut, cre] (<https://orcid.org/0000-0002-8215-3013>)
#> 2 Valerio Gherardi [aut, cre] (<https://orcid.org/0000-0002-8215-3013>)
#> 3 Valerio Gherardi [aut, cre] (<https://orcid.org/0000-0002-8215-3013>)
#> 4 Valerio Gherardi [aut, cre] (<https://orcid.org/0000-0002-8215-3013>)
#> 5                                                      Valerio Gherardi
#> 6                                                      Valerio Gherardi
#> 7 Valerio Gherardi [aut, cre] (<https://orcid.org/0000-0002-8215-3013>)
#>                             Maintainer    _user _type
#> 1 Valerio Gherardi <vgherard@sissa.it> vgherard   src
#> 2 Valerio Gherardi <vgherard@sissa.it> vgherard   src
#> 3 Valerio Gherardi <vgherard@sissa.it> vgherard   src
#> 4 Valerio Gherardi <vgherard@sissa.it> vgherard   src
#> 5 Valerio Gherardi <vgherard@sissa.it> vgherard   src
#> 6 Valerio Gherardi <vgherard@sissa.it> vgherard   src
#> 7 Valerio Gherardi <vgherard@sissa.it> vgherard   src
#>                        _file               _published
#> 1    runiv_0.0.0.9000.tar.gz 2021-07-27T15:29:39.257Z
#> 2           r2r_0.1.1.tar.gz 2021-07-25T15:23:20.981Z
#> 3        kgrams_0.1.0.tar.gz 2021-07-11T05:26:27.299Z
#> 4 scribblr_0.2.0.9000.tar.gz 2021-07-16T17:28:16.646Z
#> 5       gsample_0.1.0.tar.gz 2021-07-19T05:50:02.640Z
#> 6           sbo_0.5.0.tar.gz 2021-07-03T08:51:46.882Z
#> 7          fcci_1.0.0.tar.gz 2021-07-03T09:23:41.181Z
#>                      _builder.upstream _builder.date
#> 1    https://github.com/vgherard/runiv    1627399779
#> 2      https://github.com/vgherard/r2r    1627226600
#> 3   https://github.com/vgherard/kgrams    1625981187
#> 4 https://github.com/vgherard/scribblr    1626456496
#> 5  https://github.com/vgherard/gsample    1626673802
#> 6      https://github.com/vgherard/sbo    1625302306
#> 7     https://github.com/vgherard/fcci    1625304221
#>                            _builder.commit _builder.registered
#> 1 00e5838d42c0ff8ccdd25342332ccd903bd6b765                true
#> 2 d9679b09071a7d71bb06924850d6566c8582e99b                true
#> 3 3760408b9d0eafd4ff826201c9a5576f57a25500                true
#> 4 7ae46bf068533d6a9956d97ab89a69b2ba6471f4                true
#> 5 5b1bbc8e768417389b6ca1451e80f93457cbcfba                true
#> 6 75374a5bf57709fae780207fa1f93b8e86d7c6c4                true
#> 7 3df3ccab1ab9f0880c9a4140430851ff366a3669                true
#>   _builder.timestamp _builder.maintainerlogin _builder.distro  _builder.host
#> 1         1627398596                 vgherard           focal GitHub-Actions
#> 2         1625484226                 vgherard           focal GitHub-Actions
#> 3         1613042037                 vgherard           focal GitHub-Actions
#> 4         1626452731                 vgherard           focal GitHub-Actions
#> 5         1611175216                 vgherard           focal GitHub-Actions
#> 6         1607180351                 vgherard           focal GitHub-Actions
#> 7         1617373968                 vgherard           focal GitHub-Actions
#>   _builder.status
#> 1         success
#> 2         success
#> 3         success
#> 4         success
#> 5         success
#> 6         success
#> 7         success
#>                                                     _builder.url
#> 1 https://github.com/r-universe/vgherard/actions/runs/1071696929
#> 2 https://github.com/r-universe/vgherard/actions/runs/1064845553
#> 3 https://github.com/r-universe/vgherard/actions/runs/1019368977
#> 4 https://github.com/r-universe/vgherard/actions/runs/1038259138
#> 5 https://github.com/r-universe/vgherard/actions/runs/1043988360
#> 6  https://github.com/r-universe/vgherard/actions/runs/995955933
#> 7  https://github.com/r-universe/vgherard/actions/runs/996009318
#>                                                                                                                                              _builder.vignettes
#> 1                                                                                                                                                          NULL
#> 2 benchmarks.Rmd, r2r.Rmd, benchmarks.html, r2r.html, Comparison, r2r, 2021-07-02 11:17:51, 2021-07-01 15:23:59, 2021-07-02 12:04:10, 2021-07-02 11:17:51, 3, 2
#> 3                                                                kgrams.Rmd, kgrams.html, kgrams, Valerio Gherardi, 2021-02-06 18:04:34, 2021-02-11 11:13:57, 4
#> 4                                                                                                                                                          NULL
#> 5                                                                                                                                                          NULL
#> 6                                                                        sbo.Rmd, sbo.html, sbo, Valerio Gherardi, 2020-08-01 22:17:39, 2020-12-05 11:01:42, 10
#> 7                                                                                                                                                          NULL
#>                    _builder.sysdeps
#> 1                              NULL
#> 2                              NULL
#> 3 libstdc++6, 10.2.0-5ubuntu1~20.04
#> 4                              NULL
#> 5 libstdc++6, 10.2.0-5ubuntu1~20.04
#> 6 libstdc++6, 10.2.0-5ubuntu1~20.04
#> 7 libstdc++6, 10.2.0-5ubuntu1~20.04
#>                                                     _builder.pkglogo
#> 1                                                               <NA>
#> 2                                                               <NA>
#> 3                                                               <NA>
#> 4 https://github.com/vgherard/scribblr/raw/HEAD/man/figures/logo.png
#> 5                                                               <NA>
#> 6                                                               <NA>
#> 7                                                               <NA>
#>                             MD5sum
#> 1 0cc79b701cd26f350c3c631a438bfbe4
#> 2 6250f28963e5cf34914a83c4a290bad1
#> 3 c20d7103b32574373cb0a576e3be6c96
#> 4 9bc8757ade7e344ee9b1c955dd3b7fd3
#> 5 8340aad7192cd6b7c5d04565e9ac00af
#> 6 acd1abeb9b46a1c607d9f60aea0b8510
#> 7 9ca24cd15c82ac543d5f205096d5eba4
#>                                                                                                                                                                                                                                            _hard_deps
#> 1                                                                                                                                                                                     curl, httr, jsonlite, rlang, Imports, Imports, Imports, Imports
#> 2                                                                                                                                                                                                                     digest, rlang, Imports, Imports
#> 3                                           Rcpp, RcppProgress, Rcpp, rlang, methods, utils, RcppProgress, Rdpack, R, LinkingTo, LinkingTo, Imports, Imports, Imports, Imports, Imports, Imports, Depends, NA, NA, NA, NA, NA, NA, >= 0.1, NA, >= 3.5
#> 4 shiny, miniUI, shinyjs, rstudioapi, fs, usethis, keys, gh, shinyWidgets, utils, rlang, >= 0.13, >= 0.1.1, NA, >= 0.5, NA, NA, NA, NA, NA, NA, NA, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports
#> 5                                                                                                                                                                                                                      Rcpp, Rcpp, LinkingTo, Imports
#> 6                                    R, Rcpp, testthat, Rcpp, rlang, tidyr, dplyr, utils, stats, graphics, >= 3.5.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, Depends, LinkingTo, LinkingTo, Imports, Imports, Imports, Imports, Imports, Imports, Imports
#> 7                                                                                                                                                                                 Rcpp, Rcpp, assertthat, rlang, LinkingTo, Imports, Imports, Imports
#>                                                                                                                                         _soft_deps
#> 1                                                                                                 covr, testthat, Suggests, Suggests, NA, >= 3.0.0
#> 2 knitr, covr, testthat, rmarkdown, microbenchmark, hash, Suggests, Suggests, Suggests, Suggests, Suggests, Suggests, NA, NA, >= 3.0.0, NA, NA, NA
#> 3                                                   testthat, covr, knitr, rmarkdown, >= 3.0.0, NA, NA, NA, Suggests, Suggests, Suggests, Suggests
#> 4                                                                                                 covr, testthat, Suggests, Suggests, NA, >= 3.0.0
#> 5                                                                                         knitr, rmarkdown, testthat, Suggests, Suggests, Suggests
#> 6                                       ggplot2, knitr, rmarkdown, cli, testthat, covr, Suggests, Suggests, Suggests, Suggests, Suggests, Suggests
#> 7                                                                                                 covr, testthat, Suggests, Suggests, NA, >= 3.0.0
#>   VignetteBuilder LazyData SystemRequirements RdMacros    Type
#> 1            <NA>     <NA>               <NA>     <NA>    <NA>
#> 2           knitr     <NA>               <NA>     <NA>    <NA>
#> 3           knitr     true              C++11   Rdpack    <NA>
#> 4            <NA>     <NA>               <NA>     <NA>    <NA>
#> 5           knitr     <NA>              C++11     <NA> Package
#> 6           knitr     true              C++11     <NA> Package
#> 7            <NA>     <NA>              C++11     <NA> Package